mcp/pyproject.toml
[project]
name = "databricks-mcp"
version = "0.0.1"
description = "MCP helpers for Databricks"
authors = [
    { name="Databricks", email="agent-feedback@databricks.com" },
]
readme = "README.md"
license = { text="Apache-2.0" }
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.9.1",
    "databricks-sdk>=0.49.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "typing_extensions",
  "databricks-sdk>=0.49.0",
  "ruff==0.6.4",
]

integration = [
  "pytest-timeout>=2.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
  "src/databricks_mcp/*"
]

[tool.hatch.build.targets.wheel]
packages = ["src/databricks_mcp"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
  # isort
  "I",
  # bugbear rules
  "B",
  # remove unused imports
  "F401",
  # bare except statements
  "E722",
  # print statements
  "T201",
  "T203",
  # misuse of typing.TYPE_CHECKING
  "TCH004",
  # import rules
  "TID251",
  # undefined-local-with-import-star
  "F403",
]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 88

[tool.ruff.lint.pydocstyle]
convention = "google"
